# WebAPI/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MyAppAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Venue": {
      "get": {
        "tags": [
          "Venue"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Venue"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Venue"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Venue"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Venue"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Venue/{id}": {
      "get": {
        "tags": [
          "Venue"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "City": {
      "type": "object",
      "properties": {
        "cityId": {
          "format": "int32",
          "type": "integer"
        },
        "code": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "name": {
          "maxLength": 250,
          "minLength": 0,
          "type": "string"
        }
      }
    },
...
    "Venue": {
      "required": [
        "cityId",
        "isActive"
      ],
      "type": "object",
      "properties": {
        "venueId": {
          "format": "int32",
          "type": "integer"
        },
        "code": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "name": {
          "maxLength": 250,
          "minLength": 0,
          "type": "string"
        },
        "latLong": {
          "maxLength": 250,
          "minLength": 0,
          "type": "string"
        },
        "contactNumber": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "address": {
          "maxLength": 250,
          "minLength": 0,
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "city": {
          "$ref": "#/definitions/City"
        },
        "cityId": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}